Account Functionality
Change Address fully functional.
Change Password fully functional.
Fixed typo in Activity_Password_Popup.
Can't work on username and email due to lack of proper TextViews.
AccountFragment is a mess with repeated code, but it works, so lets leave it.
If you're reading this, know that I'm pushing this at 03:33 and my brain is no longer able to can.